Civic_Access/backend/requirements.txt
fastapi==0.123.9
uvicorn==0.38.0
sqlalchemy==2.0.44
python-dotenv==1.0.1
bcrypt==5.0.0
PyJWT==2.10.1
pydantic==2.12.4
pydantic-settings==2.5.0
python-multipart==0.0.7
fastapi-cors==0.0.6
email-validator==2.3.0